+++ packages/core/src/state/index.ts
@@ -54,8 +54,8 @@ export function createRouteState(readyTimeout: number = DEFAULT_READY_TIMEOUT) {
 		currentRoute = createRouteLocation({
 			path: route.path,
 			name: route.name,
-			meta: Object.freeze({ ...route.meta }),
-			query: Object.freeze({ ...route.query }),
+			meta: { ...route.meta },
+			query: { ...route.query },
 			fullPath: route.fullPath,
 			params: route.params,
 			...(route._synced !== undefined && { _synced: route._synced })